+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 2.5.1
 * Added consumable items to CreatureConfig
 * Added various properties to CreatureConfig
+* Added cumulative level effects for custom creatures (thx A Sharp Pen)
 
 ## Version 2.5.0
 * Added utility methods for texture rescaling to Utils.ShaderHelper